+++ Admin.Persistence/Repositories/Base/BaseLongRepository.cs
@@ -21,7 +21,7 @@ public class BaseLongRepository<T> : IBaseLongRepository<T> where T : BaseLongEn
         return await _context.SaveChangesAsync() > 0;
     }
 
-    public async Task<T> SelectByPk(int id)
+    public async Task<T> SelectByPk(long id)
     {
         return await _dbSet.FindAsync(id) ?? throw new InvalidOperationException($"Não encontrado {typeof(T).Name} id:{id}");
     }
